Display empty buglists with number zero.
This is useful because the buglist summaries do not show the count of
security bugs, so it is useful to be able to click on an "empty" list
that might actually contain some security bugs. Without this change, the
user has to hunt for the nbsp, which is difficult and counter-intuitive.